install-debian: Rewrite the installer as a shell script
The original proof-of-concept installer was written in make. This was
great for initial hacking but imposes limits on us that are starting
to show.

The shell script is more-or-less just a transliteration of the make
rules into shell syntax plus a bit of glue for command line processing.

Whilst rewriting the installer we also introduce template processing and
use this to avoid problems when using the installer from one of its
own installations (see BLKNAME in README.md).