install-debian: Add RELEASE= support
Currently this only allows us to choose between buster and bullseye
but it should be a simple addition to add support for sid now.

We also introduce DEBOOTSTRAP_PACKAGES to force additional packages
into the debootstrap tarball. This is mostly a testing option as it
allows us to cache all the GUI packages before we launch the task
select.