Accept *mmc suffixes on the default device path.
On my system (Debian testing/unstable with the danielt platform-
specific 5.5 kernel) the external sdcard slot posts as
`/dev/disk/by-path/platform-fe320000.mmc` rather than `.dwmmc`,
so the default doesn't work. Users are thus encouraged to set
the BLKDEV directly which adds to the difficulty of the
installer.

Instead, accept any suffix ending in `mmc` so that both styles
are covered. This would also match a hypothetical `emmc` suffix,
but the `platform-fe32` prefix should ensure that the default
still refers to the more-easily repaired external card.