Support RELEASE=sid/unstable.
Add sources.list options for the Debian unstable channel. Note
that these can't be derived from a template shared with the
buster and bullseye targets, because unstable doesn't have
a separate security or updates package repository.

Since builds of the device-specific kernel are only available
for buster and bullseye, fall back to the bullseye kernel
repository when the target release doesn't have a matching
configuration.

With these changes, passing `RELEASE=unstable` successfully
creates a bootable image, saving the migration step for users
who want to follow that repo.